@@ -1,21 +1,17 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.css';
-import reportWebVitals from './reportWebVitals';
 import App from './App';
-import {Provider} from 'react-redux';
-import {store} from './state/store';
+import * as serviceWorker from './serviceWorker';
+import { createRoot } from 'react-dom/client';
+import {Provider} from "react-redux";
+import {store} from "./store/store";
 
-ReactDOM.render(
-    <React.StrictMode>
-        <Provider store={store}>
-            <App/>
-        </Provider>
-    </React.StrictMode>,
-    document.getElementById('root')
-);
+const container  = document.getElementById('root') as HTMLElement
+const root = createRoot(container);
+root.render(<Provider store={store}><App /></Provider>);
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: https://bit.ly/CRA-PWA
+serviceWorker.unregister();
+
